Role value should be changed.
Currently role="dialog" causes an issue for screen readers like NVDA or Jaws to read the content that is inside of a modal. Proposing the value of role to be changed to role="content" so that applications like NVDA or JAWS will automatically start reading the contents of a modal once the modal opens.